veyron_bot/requirements.txt
tweepy>=4.14.0
anthropic>=0.40.0
python-dotenv>=1.0.0
schedule>=1.2.0
